requirements.txt
Django==1.10.3
Pillow==3.4.2
requests==2.12.4
selenium==3.0.2
beautifulsoup4==4.5.1
